Add max-col-width functional test
