--- src/components/ids-layout-grid/demos/homepages-example-three.html
@@ -1,123 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <title><%= htmlWebpackPlugin.options.title %></title>
-    <%= htmlWebpackPlugin.options.font %>
-  </head>
-  <body>
-    <ids-container role="main" padding="0" hidden>
-      <ids-masthead title="Infor Application" icon="logo" role="navigation">
-        <section title="start" slot="start">
-          <ids-button icon="grid"><span slot="text" class="audible">Grid Button</span></ids-button>
-        </section>
-        <section title="center" slot="center">
-          <ids-button icon="home"><span slot="text">Home</span></ids-button>
-          <ids-button icon="star-outlined"><span slot="text">Star</span></ids-button>
-          <ids-button icon="info"><span slot="text">Info</span></ids-button>
-        </section>
-        <section title="end" slot="end">
-          <ids-button icon="user"><span slot="text" class="audible">User Button</span></ids-button>
-          <ids-button icon="mingle-share"><span slot="text" class="audible">Mingle Button</span></ids-button>
-          <ids-button icon="bookmark-outlined"><span slot="text" class="audible">Bookmark Button</span></ids-button>
-        </section>
-      </ids-masthead>
-      <ids-header>
-        <ids-toolbar>
-          <ids-toolbar-section type="title">
-            <ids-text font-size="20">My Header</ids-text>
-            <ids-text font-size="14">With some extra information below</ids-text>
-          </ids-toolbar-section>
-          <ids-toolbar-section align="end">
-            <ids-search-field id="header-search" clearable [attr.color-variant]="'alternate'" [attr.label-state]="'collapsed'" no-margins></ids-search-field>
-          </ids-toolbar-section>
-        </ids-toolbar>
-      </ids-header>
-
-      <section class="section-container">
-        <ids-layout-grid
-          auto="true"
-          gap="lg"
-          min-col-width="360px"
-          max-col-width="360px"
-          auto-flow="dense"
-          auto-rows="1fr"
-          justify-content="center"
-        >
-          <ids-layout-grid-cell>
-            <ids-card>
-              <div slot="card-header">
-                <ids-text font-size="20" type="h2" overflow="ellipsis" tooltip="true">Widget Title 1</ids-text>
-              </div>
-              <div slot="card-content">
-              </div>
-            </ids-card>
-          </ids-layout-grid-cell>
-          <ids-layout-grid-cell>
-            <ids-card>
-              <div slot="card-header">
-                <ids-text font-size="20" type="h2" overflow="ellipsis" tooltip="true">Widget Title 2</ids-text>
-              </div>
-              <div slot="card-content">
-              </div>
-            </ids-card>
-          </ids-layout-grid-cell>
-          <ids-layout-grid-cell>
-            <ids-card>
-              <div slot="card-header">
-                <ids-text font-size="20" type="h2" overflow="ellipsis" tooltip="true">Widget Title 3</ids-text>
-              </div>
-              <div slot="card-content">
-              </div>
-            </ids-card>
-          </ids-layout-grid-cell>
-          <ids-layout-grid-cell>
-            <ids-card auto-fit="true">
-              <div slot="card-header">
-                <ids-text font-size="20" type="h2" overflow="ellipsis" tooltip="true">Widget Title 4</ids-text>
-              </div>
-              <div slot="card-content">
-              </div>
-            </ids-card>
-          </ids-layout-grid-cell>
-          <ids-layout-grid-cell col-span="2">
-            <ids-card>
-              <div slot="card-header">
-                <ids-text font-size="20" type="h2" overflow="ellipsis" tooltip="true">Widget Title 5</ids-text>
-              </div>
-              <div slot="card-content">
-              </div>
-            </ids-card>
-          </ids-layout-grid-cell>
-          <ids-layout-grid-cell col-span="2">
-            <ids-card>
-              <div slot="card-header">
-                <ids-text font-size="20" type="h2" overflow="ellipsis" tooltip="true">Widget Title 6</ids-text>
-              </div>
-              <div slot="card-content">
-              </div>
-            </ids-card>
-          </ids-layout-grid-cell>
-          <ids-layout-grid-cell col-span="2">
-            <ids-card>
-              <div slot="card-header">
-                <ids-text font-size="20" type="h2" overflow="ellipsis" tooltip="true">Widget Title 7</ids-text>
-              </div>
-              <div slot="card-content">
-              </div>
-            </ids-card>
-          </ids-layout-grid-cell>
-          <ids-layout-grid-cell col-span="2">
-            <ids-card>
-              <div slot="card-header">
-                <ids-text font-size="20" type="h2" overflow="ellipsis" tooltip="true">Widget Title 8</ids-text>
-              </div>
-              <div slot="card-content">
-              </div>
-            </ids-card>
-          </ids-layout-grid-cell>
-        </ids-layout-grid>
-      </section>
-
-    </ids-container>
-  </body>
-</html>
